pyproject.toml
[tool.poetry]
name = "ekklesia-portal"
version = "21.05.0"
description = "Motion portal for the Ekklesia e-democracy platform"
readme = "README.rst"
keywords = ["ekklesia", "edemocracy", "argument mapping", "evoting", "participation"]
license = "AGPL-3.0-or-later"
authors = ["Tobias dpausp <dpausp@posteo.de>"]
classifiers = [
    "License :: OSI Approved :: GNU Affero General Public License v3 or later (AGPLv3+)",
    "License :: OSI Approved :: GNU General Public License v3 (GPLv3)",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.9"
]


[tool.poetry.dependencies]
python = "^3.9"
alembic = "^1.4.2"
dataclasses-json = "^0.4.3"
passlib = "^1.7.2"
psycopg2 = "^2.8.5"
sqlalchemy-searchable = "^1.1.0"
ekklesia-common = {git = "https://github.com/edemocracy/ekklesia-common", rev = "wip"}
base32-crockford = "^0.3.0"
py-gfm = "0.1.4"
markdown = "2.6.11"
babel = "^2.9.0"
deform = "^2.0.15"
eliot-tree = "^19.0.1"
pdbpp = "^0.10.2"


[tool.poetry.dev-dependencies]
colorama = "^0.4.3"
coverage = {extras = ["toml"], version = "^5.2.1"}
mimesis-factory = "^1.1.0"
pytest = "^6.0.1"
pytest-cov = "^2.8.1"
pytest-factoryboy = "^2.0.3"
pytest-instafail = "^0.4.1"
pytest-localserver = "^0.5.0"
pytest-mock = "^3.1.0"
pytest-pspec = "^0.0.4"
responses = "^0.10.14"
typing-extensions = "^3.7.4"
webtest = "^2.0.35"


[tool.coverage.report]
skip_covered = true
# Regexes for lines to exclude from consideration
exclude_lines = [
    "pragma: no cover",
    "raise AssertionError",
    "raise NotImplementedError",
    "if __name__ == .__main__.:"
]


[tool.coverage.run]
source = [
    "ekklesia_portal"
]
omit = [
    "src/ekklesia_portal/runserver.py",
    "src/ekklesia_portal/concepts/*/__init__.py",
    "src/ekklesia_portal/runserver_production_deps.py",
    "src/ekklesia_portal/pathtool.py",
    "src/ekklesia_portal/whytool.py"
]

[tool.pylint.basic]
good-names = "q, s, f, wf, m, k, v, id"

[tool.pylint.imports] # Taken care of by isort, mostly included for error messages
known-third-party="ekklesia_common"


[tool.pylint.master]
disable = "C0114, C0115, C0116, R0903"


[tool.pylint.format]
max-line-length="120"
ignore-long-lines="^\\s*(# )?<?https?://\\S+>?$"


[tool.pytest.ini_options]
addopts = """\
  --show-capture=no \
  -ra \
  --tb=short \
  -k "not integration" \
  --strict-markers \
"""
filterwarnings = [
  "ignore::DeprecationWarning"
]
norecursedirs = "tests/helpers"
testpaths = "tests"
markers = [
  "integration: An Integration test which is allowed to modify the test database."
]


[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
